% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read.R
\name{fars_read}
\alias{fars_read}
\title{Read fars data.}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{(required) A character string specifying path to the fars
data.}
}
\value{
A tibble, i.e. an object that inherits from classes \code{"tbl_df"},
 \code{"tbl"}, and \code{"data.frame"} (for details, see \url{https://tibble.tidyverse.org/}).
}
\description{
Function that reads fars data from a file
and converts it into a tibble.
}
\examples{
\dontrun{
fars_read("fars.data.csv")
}

}
