% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read_years.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Get year / month accident incidence.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A numeric, integer, or character vector specifying the years for
which accident occurrence will be collected.}
}
\value{
A list with as many elements as there elements in the input vector
  \code{years}. Each element in the list represent on year;  it is a tible
  with column \code{"MONTH"} and \code{"year"}.
}
\description{
Function that collects accidents, and the
month that they happened, for specified years.
}
\details{
The function checks whether a file for a given year exists; if not,
  a warning is issued and \code{NULL} is returned for that year.
}
\examples{
\dontrun{
 fars_read_year(c(2013, 2014))
}

}
