% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_summarize_years.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize accidents by year and month.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A numeric, integer, or character vector specifying the years for
which accident occurrence will be collected.}
}
\value{
A tibble summarizing the number of accidents for each month
and year (months are in rows and years
in columns).
}
\description{
Function that takes as input a vector of years,
reads fars data for them, and summarizes accident
incidence by year and month.
}
\examples{
\dontrun{
far_summarize_years(c(2013, 2014))
}

}
