% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_filename.R
\name{make_filename}
\alias{make_filename}
\title{Construct file name}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A character or integer vector of length 1, specifying the year.}
}
\value{
A character vector of length 1.
}
\description{
Funtion that constructs a character string representing the filename of fars
data for a specified year.
}
\examples{
make_filename(1987)
make_filename("1987")

}
